[#26]refactor: soma.haeya.edupi_db -> soma.edupi.db
