+++ src/test/java/soma/edupi/db/EdupiUserApplicationTests.java
@@ -1,4 +1,4 @@
-package soma.haeya.edupi_db;
+package soma.edupi.db;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class EdupiUserApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
